// packages/node-webcodecs/lib/types.ts
export type AllowSharedBufferSource = ArrayBuffer | ArrayBufferView;
// ...
export type EncodedAudioChunkType = 'key' | 'delta';

/**
 * WebIDL:
 * enum EncodedVideoChunkType { "key", "delta" };
 */
export type EncodedVideoChunkType = 'key' | 'delta';
// ...
export interface EncodedVideoChunkInit {
  type: EncodedVideoChunkType;
  timestamp: number; // long long, microseconds
  duration?: number; // unsigned long long, microseconds
  data: AllowSharedBufferSource;
  transfer?: ArrayBuffer[];
}
// ...
export interface EncodedAudioChunkInit {
  type: EncodedAudioChunkType;
  timestamp: number; // long long, microseconds
  duration?: number; // unsigned long long, microseconds
  data: AllowSharedBufferSource;
  transfer?: ArrayBuffer[];
}